requirements.txt
playwright==1.40.0
flask==2.3.3
apscheduler==3.10.1
pyyaml==6.0.1
pandas==2.1.0
python-dotenv==1.0.0
requests==2.31.0
